requirements.txt
pydantic>=2.0.0,<3.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
anthropic>=0.30.0,<1.0.0
requests>=2.31.0
pytest>=8.0.0
ruff>=0.4.0
